Add test for between_ages user scope
We were already testing it as part of the statisticable concern, but
it's easier to understand when we have specific tests for it as well.

Note that we're using `eq` insteab of `be` because it's what we use most
often in the test. For consistency, we're also changing the tesst in
budget stats so it uses `eq`.