Remove obsolete routes for poll questions
The routes for poll questions were accidentally deleted in commit
5bb831e9595 when deleting the `:show` action, and restored in commit
9871503c5e8. However, the deleted code was:

```
resources :questions, only: [:show], controller: 'polls/questions' (...)
```

While the restored code was:

```
resources :questions, controller: 'polls/questions' (...)
```

Meaning we forgot to add the `only: []` option when restoring the
routes.

We also forgot to remove the `before_action` code when deleting the
`:show` action, so we're removing it now.