Use one SUMMARY.md file per language
On GitBook, currently Spanish and English belong in different spaces, so
we're going to provide two different github integrations (one per
language).

So we're using a different SUMMARY.md file per language.

For consistency, since there was a link to the README file in the
Spanish summary, we're also adding it to the English one. It was working
in any case because on our current gitbook page the summary was
generated through gitbook.

Since we no longer have any of the files referenced in the
`.gitbook.yaml` file, we're removing it as well.

Note that there was a typo in the English summary, which pointed to the
Spanish license. It didn't matter, though, since the license file is
actually in English in both cases.