Don't run linters in Code Climate
We're now running the linters in github actions, and code climate
doesn't use the same versions of the gems or npm packages that we're
using in the application. For example, only a few limited rubocop
versions are available, and sometimes code climate breaks because our
rubocop version isn't compatible with theirs. Now we're also using a
version of stylelint they don't support.

All in all, the linters check in code climate isn't helpful at all
anymmore.