Add a more specific link for local installation docs
On our regular README, this link used to point to the documentation
site, which made sense because the documentation used to be on a
different repository. Now that the documentation is on this repository
as well, we can directly link to that file.

And, on the documentation README, we can also link to the local
installation instructions, just like we do in other places.