Check different content in select investments test
Since we were checking something that was already there, if the form
were submitted using an AJAX request (which is the default with Turbo,
although we don't use it yet), there would be a chance that th request
doesn't finish before the test does, leading to potential failures in
the next test.

This way we're also checking that the "Filter" button actually does
something when selecting the "Selected" filter.