Update sassc-embedded to 1.80.4 to fix foundation 6.9.0 precompile error
When updating to foundation-sites 6.9.0, the assets:precompile task fails
due to the use of the color.channel method in the smart-scale function,
introduced in foundation 6.9.0.

This method is only available in sass-embedded from version 1.79.4 onwards [1].
The current setup uses sassc-embedded 1.70.1 and sass-embedded 1.72.0,
which lack this functionality.

We could simply run a bundle update sassc-embedded, which would automatically
update the sass-embedded dependency to a compatible version (>= 1.79.4),
resolving the issue. However, to ensure both gems stay aligned and to avoid
potential future incompatibilities, we will explicitly update the sassc-embedded
version in the Gemfile to ~> 1.80.1.

[1] https://rubydoc.info/gems/sass-embedded/1.79.4/Sass/Value/Color:channel