Add missing expectations in users tests
One of these tests has failed in our CI with the following message:

```
1) Users Public activity user can hide public page

   Failure/Error: expect(page).to have_content("activity list private")
     expected to find text "activity list private" in "Language: \n
     \nEnglish\nDeutsch\nEspañol\nFrançais\nNederlands\nPortuguês
     brasileiro\n中文\n       Notifications\nYou are in\nMy content\nMy
     account\nSign out\nDebates\nProposals\nVoting\nCollaborative
     legislation\nParticipatory budgeting\nSDG\nHelp\nM\nManuela124\nUser has
     no public activity\nOpen government\nThis portal uses the CONSUL
     DEMOCRACY application which is open-source
     software.\nParticipation\nDecide how to shape the city you want to live
     in.\nCONSUL DEMOCRACY, 2024 Privacy Policy Terms and conditions of use
     Accessibility"
```

Note how the text "User has no public activity" is present, which is a
message that appears when the user's activity is public.

A possible explanation is that we didn't check that the request done by
the "Save changes" button had finished before continuing with the tests.
Back when we wrote this test, submitting a form in a test would always
wait for the request to be finished before continuing, but a lot has
changed since then.

So we're adding an expectation to make sure the the changes have been
saved before making a new request.

We're also rearraging the blank lines in these tests and removing the
parenthesis in `have_content` expectations to be consistent with the
expectations we're adding.